backend/requirements.txt
langchain
langchain-community
langchain-google-genai
langgraph
python-dotenv
google-search-results
fastapi[standard]
langchain-google-vertexai
faiss-cpu
python-pptx
flask
Flask-JWT-Extended
scrypt
elasticsearch==8.17.0
azure-storage-blob==12.24.0
python-logstash==0.4.8
flask
Flask-JWT-Extended
scrypt
bas4
faiss-cpu
